backend/pyproject.toml
[project]
name = "book-club-backend"
version = "0.0.0"
description = "Book Club backend — FastAPI modular monolith"
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = { text = "Proprietary" }
authors = [{ name = "Book Club Team" }]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pyjwt[crypto]>=2.9",
    "boto3>=1.35",
    "httpx>=0.27",
    "tenacity>=9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "testcontainers[postgres]>=4.8",
    "respx>=0.21",
    "types-boto3>=1.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501", # handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["N802", "N803"]
"migrations/**" = ["N999"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["boto3.*", "botocore.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
